adcom1: int type corrections: utcoffset - int, w/h ratios - uint8

+++ adcom1/display.go
@@ -58,7 +58,9 @@ type Display struct {
 	// Definition:
 	//   Relative width of the creative when expressing size as a ratio, typically for non-native ads.
 	//   Note that mixing absolute and relative sizes is not recommended.
-	WRatio uint64 `json:"wratio,omitempty"`
+	// Dev note:
+	//   This is kept as `uint8` because ratio values are expected to be quite small (like 16:9).
+	WRatio uint8 `json:"wratio,omitempty"`
 
 	// Attribute:
 	//   hratio
@@ -67,7 +69,9 @@ type Display struct {
 	// Definition:
 	//   Relative height of the creative when expressing size as a ratio, typically for non-native ads.
 	//   Note that mixing absolute and relative sizes is not recommended.
-	HRatio uint64 `json:"hratio,omitempty"`
+	// Dev note:
+	//   This is kept as `uint8` because ratio values are expected to be quite small (like 16:9).
+	HRatio uint8 `json:"hratio,omitempty"`
 
 	// Attribute:
 	//   priv
