+++ adcom1/geo.go
@@ -113,7 +113,9 @@ type Geo struct {
 	//   integer
 	// Definition:
 	//   Local time as the number +/- of minutes from UTC.
-	UTCOffset int64 `json:"utcoffset,omitempty"`
+	// Dev note:
+	//   This field is kept as `int` to follow type choice for timezone offset in std. `time` package.
+	UTCOffset int `json:"utcoffset,omitempty"`
 
 	// Attribute:
 	//   ext
